resoolving conflicts with gitignore and extras
