+++ .gitignore
@@ -5,3 +5,6 @@ __pycache__/
 db
 venv
 .vscode
+db.sqlite3
+.env.db
+.vscode/
